Update ViennaCL dependency to sources only.
Until now, the OpenCL packages depended on a full build of the ViennaCL 1.7.1
library. In fact, it is sufficient to only have ViennaCL sources. This speeds
up installation and avoids irrelevant build problems. (For example, gcc 5.4.0
dies with an internal compiler error when building Caffe with ViennaCL on
Odroid XU3.)